работает отображение сотрудников без кнопок пока

+++ birthday_congrats/databases/sql/sessions.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS `sessions`;
 CREATE TABLE `sessions` (
   `sess_id` text NOT NULL,
-  `user_id` int(32) NOT NULL,
-  `expires` int(64)
+  `user_id` int NOT NULL,
+  `expires` bigint
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
